Changed int64 to int.

+++ manager/memory/manager_memory.go
@@ -52,7 +52,7 @@ func (m *MemoryManager) Update(policy Policy) error {
 }
 
 // GetAll returns all policies.
-func (m *MemoryManager) GetAll(limit, offset int64) (Policies, error) {
+func (m *MemoryManager) GetAll(limit, offset int) (Policies, error) {
 	keys := make([]string, len(m.Policies))
 	i := 0
 	m.RLock()
@@ -128,7 +128,7 @@ func (m *MemoryManager) FindRequestCandidates(r *Request) (Policies, error) {
 // FindPoliciesForSubject returns policies that could match the subject. It either returns
 // a set of policies that applies to the subject, or a superset of it.
 // If an error occurs, it returns nil and the error.
-func (m *MemoryManager) FindPoliciesForSubject(subject string, limit int64, offset int64) (Policies, error) {
+func (m *MemoryManager) FindPoliciesForSubject(subject string, limit int, offset int) (Policies, error) {
 	return m.findAllPolicies()
 }
